Restrcture translationlayer tests
