+++ test/Microsoft.TestPlatform.AcceptanceTests/TranslationLayerTests/LiveUnitTestingTests.cs
@@ -3,29 +3,31 @@
 
 namespace Microsoft.TestPlatform.AcceptanceTests.TranslationLayerTests
 {
+    using System;
     using System.Collections.Generic;
     using System.Linq;
 
+    using Microsoft.TestPlatform.TestUtilities;
     using Microsoft.TestPlatform.VsTestConsole.TranslationLayer.Interfaces;
     using Microsoft.VisualStudio.TestPlatform.Common.Telemetry;
-    using Microsoft.VisualStudio.TestPlatform.ObjectModel;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel.Client;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
 
     [TestClass]
-    public class RunSelectedTests : AcceptanceTestBase
+    public class LiveUnitTestingTests : AcceptanceTestBase
     {
         private const string Netcoreapp = "netcoreapp";
         private const string Message = "VsTestConsoleWrapper donot support .Net Core Runner";
-        private IVsTestConsoleWrapper vstestConsoleWrapper;
-        private RunEventHandler runEventHandler;
-        private DiscoveryEventHandler discoveryEventHandler;
 
-        private void Setup()
+        private IVsTestConsoleWrapper vstestConsoleWrapper;
+        private DiscoveryEventHandler discoveryEventHandler;
+        private DiscoveryEventHandler2 discoveryEventHandler2;
+
+        public void Setup()
         {
             this.vstestConsoleWrapper = this.GetVsTestConsoleWrapper();
-            this.runEventHandler = new RunEventHandler();
             this.discoveryEventHandler = new DiscoveryEventHandler();
+            this.discoveryEventHandler2 = new DiscoveryEventHandler2();
         }
 
         [TestCleanup]
@@ -34,53 +36,60 @@ namespace Microsoft.TestPlatform.AcceptanceTests.TranslationLayerTests
             this.vstestConsoleWrapper?.EndSession();
         }
 
+
         [TestMethod]
         [NetFullTargetFrameworkDataSource]
         [NetCoreTargetFrameworkDataSource]
-        public void RunSelectedTestsWithoutTestPlatformOptions(RunnerInfo runnerInfo)
+        public void DiscoverTestsUsingLiveUnitTesting(RunnerInfo runnerInfo)
         {
             AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
             this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
             this.Setup();
 
-            this.vstestConsoleWrapper.DiscoverTests(this.GetTestAssemblies(), this.GetDefaultRunSettings(), this.discoveryEventHandler);
-            var testCases = this.discoveryEventHandler.DiscoveredTestCases;
+            string runSettingsXml = @"<?xml version=""1.0"" encoding=""utf-8""?>
+                                    <RunSettings>
+                                        <RunConfiguration>
+                                        <DisableAppDomain>true</DisableAppDomain>
+                                        <DisableParallelization>true</DisableParallelization>
+                                        </RunConfiguration>
+                                    </RunSettings>";
 
-            this.vstestConsoleWrapper.RunTests(testCases, this.GetDefaultRunSettings(), this.runEventHandler);
+            this.vstestConsoleWrapper.DiscoverTests(
+               this.GetTestAssemblies(),
+                runSettingsXml,
+                this.discoveryEventHandler);
+
+            // Assert
+            Assert.AreEqual(6, this.discoveryEventHandler.DiscoveredTestCases.Count);
+        }
+
+        [TestMethod]
+        [NetFullTargetFrameworkDataSource]
+        [NetCoreTargetFrameworkDataSource]
+        public void RunTestsWithLiveUnitTesting(RunnerInfo runnerInfo)
+        {
+            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
+            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
+            this.Setup();
+
+            string runSettingsXml = @"<?xml version=""1.0"" encoding=""utf-8""?>
+                                    <RunSettings>
+                                        <RunConfiguration>
+                                        <DisableAppDomain>true</DisableAppDomain>
+                                        <DisableParallelization>true</DisableParallelization>
+                                        </RunConfiguration>
+                                    </RunSettings>";
+
+            this.vstestConsoleWrapper.RunTests(
+                this.GetTestAssemblies(),
+                runSettingsXml,
+                this.runEventHandler);
 
             // Assert
             Assert.AreEqual(6, this.runEventHandler.TestResults.Count);
             Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
             Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Failed));
             Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Skipped));
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        public void RunSelectedTestsWithTestPlatformOptions(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-            this.Setup();
-
-            this.vstestConsoleWrapper.DiscoverTests(this.GetTestAssemblies(), this.GetDefaultRunSettings(), this.discoveryEventHandler);
-            var testCases = this.discoveryEventHandler.DiscoveredTestCases;
-
-            this.vstestConsoleWrapper.RunTests(
-                testCases,
-                this.GetDefaultRunSettings(),
-                new TestPlatformOptions() { CollectMetrics = true },
-                this.runEventHandler);
-
-            // Assert
-            Assert.AreEqual(6, this.runEventHandler.TestResults.Count);
-            Assert.IsTrue(this.runEventHandler.Metrics.ContainsKey(TelemetryDataConstants.TargetDevice));
-            Assert.IsTrue(this.runEventHandler.Metrics.ContainsKey(TelemetryDataConstants.TargetFramework));
-            Assert.IsTrue(this.runEventHandler.Metrics.ContainsKey(TelemetryDataConstants.TargetOS));
-            Assert.IsTrue(this.runEventHandler.Metrics.ContainsKey(TelemetryDataConstants.TimeTakenInSecForRun));
-            Assert.IsTrue(this.runEventHandler.Metrics.ContainsKey(TelemetryDataConstants.NumberOfAdapterDiscoveredDuringExecution));
-            Assert.IsTrue(this.runEventHandler.Metrics.ContainsKey(TelemetryDataConstants.RunState));
         }
 
         private IList<string> GetTestAssemblies()
