+++ test/Microsoft.TestPlatform.AcceptanceTests/TranslationLayerTests/RunTests.cs
@@ -32,10 +32,16 @@ namespace Microsoft.TestPlatform.AcceptanceTests.TranslationLayerTests
         private IVsTestConsoleWrapper vstestConsoleWrapper;
         private RunEventHandler runEventHandler;
 
-        public RunTests()
+        private void Setup()
         {
             this.vstestConsoleWrapper = this.GetVsTestConsoleWrapper();
             this.runEventHandler = new RunEventHandler();
+        }
+
+        [TestCleanup]
+        public void Cleanup()
+        {
+            this.vstestConsoleWrapper?.EndSession();
         }
 
         [TestMethod]
@@ -45,6 +51,7 @@ namespace Microsoft.TestPlatform.AcceptanceTests.TranslationLayerTests
         {
             AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
             this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
+            this.Setup();
 
             this.vstestConsoleWrapper.RunTests(this.GetTestAssemblies(), this.GetDefaultRunSettings(), this.runEventHandler);
 
@@ -62,6 +69,7 @@ namespace Microsoft.TestPlatform.AcceptanceTests.TranslationLayerTests
         {
             AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
             this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
+            this.Setup();
 
             this.vstestConsoleWrapper.RunTests(
                 this.GetTestAssemblies(),
@@ -86,6 +94,7 @@ namespace Microsoft.TestPlatform.AcceptanceTests.TranslationLayerTests
         {
             AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
             this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
+            this.Setup();
 
             this.vstestConsoleWrapper.RunTests(
                 this.GetTestAssemblies(),
@@ -101,426 +110,11 @@ namespace Microsoft.TestPlatform.AcceptanceTests.TranslationLayerTests
         [TestMethod]
         [NetFullTargetFrameworkDataSource]
         [NetCoreTargetFrameworkDataSource]
-        public void RunTestsWithTestCaseFilter(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            var sources = new List<string>
-                              {
-                                  this.GetAssetFullPath("SimpleTestProject.dll")
-                              };
-
-            this.vstestConsoleWrapper.RunTests(
-                sources,
-                this.GetDefaultRunSettings(),
-                new TestPlatformOptions() { TestCaseFilter = "FullyQualifiedName=SampleUnitTestProject.UnitTest1.PassingTest" },
-                this.runEventHandler);
-
-            // Assert
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(TestOutcome.Passed, this.runEventHandler.TestResults.FirstOrDefault().Outcome);
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        public void RunTestsWithFastFilter(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            var sources = new List<string>
-                              {
-                                  this.GetAssetFullPath("SimpleTestProject.dll")
-                              };
-
-            this.vstestConsoleWrapper.RunTests(
-                sources,
-                this.GetDefaultRunSettings(),
-                new TestPlatformOptions() { TestCaseFilter = "FullyQualifiedName=SampleUnitTestProject.UnitTest1.PassingTest | FullyQualifiedName=SampleUnitTestProject.UnitTest1.FailingTest" },
-                this.runEventHandler);
-
-            // Assert
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Failed));
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        public void RunTestsWithSourceNavigation(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            var sources = new List<string>
-                              {
-                                  this.GetAssetFullPath("SimpleTestProject.dll")
-                              };
-
-            this.vstestConsoleWrapper.RunTests(
-                sources,
-                this.GetDefaultRunSettings(),
-                new TestPlatformOptions() { TestCaseFilter = "FullyQualifiedName=SampleUnitTestProject.UnitTest1.PassingTest" },
-                this.runEventHandler);
-
-            // Assert
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(TestOutcome.Passed, this.runEventHandler.TestResults.FirstOrDefault().Outcome);
-
-            // Release builds optimize code, hence line numbers are different.
-            if (IntegrationTestEnvironment.BuildConfiguration.StartsWith("release", StringComparison.OrdinalIgnoreCase))
-            {
-                Assert.AreEqual(23, this.runEventHandler.TestResults.FirstOrDefault().TestCase.LineNumber);
-            }
-            else
-            {
-                Assert.AreEqual(22, this.runEventHandler.TestResults.FirstOrDefault().TestCase.LineNumber);
-            }
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        public void RunTestsWithTestAdapterPath(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            var testAdapterPath = Directory.EnumerateFiles(this.GetTestAdapterPath(), "*.TestAdapter.dll").ToList();
-            this.vstestConsoleWrapper.InitializeExtensions(new List<string>() { testAdapterPath.FirstOrDefault() });
-
-            this.vstestConsoleWrapper.RunTests(
-                this.GetTestAssemblies(),
-                this.GetDefaultRunSettings(),
-                this.runEventHandler);
-
-            // Assert
-            Assert.AreEqual(6, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Failed));
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Skipped));
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        public void RunTestsWithNunitAdapter(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            var sources = new List<string>
-                              {
-                                  this.GetAssetFullPath("NUTestProject.dll")
-                              };
-
-            this.vstestConsoleWrapper.RunTests(
-                sources,
-                this.GetDefaultRunSettings(),
-                this.runEventHandler);
-
-            var testCase =
-                this.runEventHandler.TestResults.Where(tr => tr.TestCase.DisplayName.Equals("PassTestMethod1"));
-
-            // Assert
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Failed));
-
-            // Release builds optimize code, hence line numbers are different.
-            if (IntegrationTestEnvironment.BuildConfiguration.StartsWith("release", StringComparison.OrdinalIgnoreCase))
-            {
-                Assert.AreEqual(11, testCase.FirstOrDefault().TestCase.LineNumber);
-            }
-            else
-            {
-                Assert.AreEqual(10, testCase.FirstOrDefault().TestCase.LineNumber);
-            }
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        public void RunTestsWithXunitAdapter(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            // Xunit >= 2.2 won't support net451, Minimum target framework it supports is net452.
-            string testAssemblyPath = null;
-            if (this.testEnvironment.TargetFramework.Equals("net451"))
-            {
-                testAssemblyPath = testEnvironment.GetTestAsset("XUTestProject.dll", "net46");
-            }
-            else
-            {
-                testAssemblyPath = testEnvironment.GetTestAsset("XUTestProject.dll");
-            }
-
-            var sources = new List<string> { testAssemblyPath };
-            var testAdapterPath = Directory.EnumerateFiles(this.GetTestAdapterPath(UnitTestFramework.XUnit), "*.TestAdapter.dll").ToList();
-            this.vstestConsoleWrapper.InitializeExtensions(new List<string>() { testAdapterPath.FirstOrDefault() });
-
-            this.vstestConsoleWrapper.RunTests(
-                sources,
-                this.GetDefaultRunSettings(),
-                this.runEventHandler);
-
-            var testCase =
-                this.runEventHandler.TestResults.Where(tr => tr.TestCase.DisplayName.Equals("xUnitTestProject.Class1.PassTestMethod1"));
-
-            // Assert
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Failed));
-
-            // Release builds optimize code, hence line numbers are different.
-            if (IntegrationTestEnvironment.BuildConfiguration.StartsWith("release", StringComparison.OrdinalIgnoreCase))
-            {
-                Assert.AreEqual(16, testCase.FirstOrDefault().TestCase.LineNumber);
-            }
-            else
-            {
-                Assert.AreEqual(15, testCase.FirstOrDefault().TestCase.LineNumber);
-            }
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        public void RunTestsWithChutzpahAdapter(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            var sources = new List<string>
-                              {
-                                  Path.Combine(this.testEnvironment.TestAssetsPath, "test.js")
-                              };
-
-            var testAdapterPath = Directory.EnumerateFiles(this.GetTestAdapterPath(UnitTestFramework.Chutzpah), "*.TestAdapter.dll").ToList();
-            this.vstestConsoleWrapper.InitializeExtensions(new List<string>() { testAdapterPath.FirstOrDefault() });
-
-            this.vstestConsoleWrapper.RunTests(
-                sources,
-                this.GetDefaultRunSettings(),
-                this.runEventHandler);
-
-            var testCase =
-                this.runEventHandler.TestResults.Where(tr => tr.TestCase.DisplayName.Equals("TestMethod1"));
-
-            // Assert
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Failed));
-            Assert.AreEqual(1, testCase.FirstOrDefault().TestCase.LineNumber);
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        [DoNotParallelize]
-        public void RunTestsWithRunSettingsWithParallel(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            string runSettingsXml = @"<?xml version=""1.0"" encoding=""utf-8""?> 
-                                    <RunSettings>     
-                                        <RunConfiguration>
-                                        <MaxCpuCount>2</MaxCpuCount>
-                                        </RunConfiguration>
-                                    </RunSettings>";
-
-            string testhostProcessName = string.Empty;
-            int expectedNumOfProcessCreated = 0;
-            if (this.IsDesktopTargetFramework())
-            {
-                testhostProcessName = "testhost.x86";
-                expectedNumOfProcessCreated = 2;
-            }
-            else
-            {
-                testhostProcessName = "dotnet";
-                if (this.IsDesktopRunner())
-                {
-                    expectedNumOfProcessCreated = 2;
-                }
-                else
-                {
-                    // Include launcher dotnet.exe
-                    expectedNumOfProcessCreated = 3;
-                }
-            }
-
-            var cts = new CancellationTokenSource();
-            var numOfProcessCreatedTask = NumberOfProcessLaunchedUtility.NumberOfProcessCreated(
-                cts,
-                testhostProcessName);
-
-            this.vstestConsoleWrapper.RunTests(
-                this.GetTestAssemblies(),
-                runSettingsXml,
-                this.runEventHandler);
-
-            cts.Cancel();
-
-            // Assert
-            Assert.AreEqual(6, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Failed));
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Skipped));
-            Assert.AreEqual(
-                expectedNumOfProcessCreated,
-                numOfProcessCreatedTask.Result,
-                $"Number of {testhostProcessName} process created, expected: {expectedNumOfProcessCreated} actual: {numOfProcessCreatedTask.Result}");
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        public void RunTestsWithTestSettings(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            var testsettingsFile = Path.Combine(Path.GetTempPath(), "tempsettings.testsettings");
-            string testSettingsXml = @"<?xml version=""1.0"" encoding=""utf-8""?><TestSettings></TestSettings>";
-
-            File.WriteAllText(testsettingsFile, testSettingsXml, Encoding.UTF8);
-            var runSettings = "<RunSettings><MSTest><SettingsFile>" + testsettingsFile + "</SettingsFile></MSTest></RunSettings>";
-            var sources = new List<string>
-                              {
-                                  this.GetAssetFullPath("MstestV1UnitTestProject.dll")
-                              };
-
-            this.vstestConsoleWrapper.RunTests(
-                sources,
-                runSettings,
-                this.runEventHandler);
-
-            // Assert
-            Assert.AreEqual(5, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Failed));
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Skipped));
-
-            File.Delete(testsettingsFile);
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        [DoNotParallelize]
-        public void RunTestsWithX64Source(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            var sources = new List<string>
-                              {
-                                  this.GetAssetFullPath("SimpleTestProject3.dll")
-                              };
-
-            string testhostProcessName = string.Empty;
-            int expectedNumOfProcessCreated = 0;
-            if (this.IsDesktopTargetFramework())
-            {
-                testhostProcessName = "testhost";
-                expectedNumOfProcessCreated = 1;
-            }
-            else
-            {
-                testhostProcessName = "dotnet";
-                if (this.IsDesktopRunner())
-                {
-                    expectedNumOfProcessCreated = 1;
-                }
-                else
-                {
-                    // Include launcher dotnet.exe
-                    expectedNumOfProcessCreated = 2;
-                }
-            }
-
-            var cts = new CancellationTokenSource();
-            var numOfProcessCreatedTask = NumberOfProcessLaunchedUtility.NumberOfProcessCreated(
-                cts,
-                testhostProcessName);
-
-            this.vstestConsoleWrapper.RunTests(
-                sources,
-                this.GetDefaultRunSettings(),
-                new TestPlatformOptions() { TestCaseFilter = "FullyQualifiedName = SampleUnitTestProject3.UnitTest1.WorkingDirectoryTest" },
-                this.runEventHandler);
-
-            cts.Cancel();
-
-            // Assert
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(1, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
-            Assert.AreEqual(
-                expectedNumOfProcessCreated,
-                numOfProcessCreatedTask.Result,
-                $"Number of {testhostProcessName} process created, expected: {expectedNumOfProcessCreated} actual: {numOfProcessCreatedTask.Result}");
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        public void RunTestsWithCustomTestHostLaunch(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            var customTestHostLauncher = new CustomTestHostLauncher();
-            this.vstestConsoleWrapper.RunTestsWithCustomTestHost(this.GetTestAssemblies(), this.GetDefaultRunSettings(), this.runEventHandler, customTestHostLauncher);
-
-            // Assert
-            Assert.AreEqual(6, this.runEventHandler.TestResults.Count);
-            Assert.IsTrue(customTestHostLauncher.ProcessId != -1);
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Failed));
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Skipped));
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
-        public void RunTestsWithLiveUnitTesting(RunnerInfo runnerInfo)
-        {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
-            this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
-
-            string runSettingsXml = @"<?xml version=""1.0"" encoding=""utf-8""?> 
-                                    <RunSettings>     
-                                        <RunConfiguration>
-                                        <DisableAppDomain>true</DisableAppDomain>
-                                        <DisableParallelization>true</DisableParallelization>
-                                        </RunConfiguration>
-                                    </RunSettings>";
-
-            this.vstestConsoleWrapper.RunTests(
-                this.GetTestAssemblies(),
-                runSettingsXml,
-                this.runEventHandler);
-
-            // Assert
-            Assert.AreEqual(6, this.runEventHandler.TestResults.Count);
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Passed));
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Failed));
-            Assert.AreEqual(2, this.runEventHandler.TestResults.Count(t => t.Outcome == TestOutcome.Skipped));
-        }
-
-        [TestMethod]
-        [NetFullTargetFrameworkDataSource]
-        [NetCoreTargetFrameworkDataSource]
         public void RunTestsShouldThrowOnStackOverflowException(RunnerInfo runnerInfo)
         {
             AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerInfo);
             this.ExecuteNotSupportedRunnerFrameworkTests(runnerInfo.RunnerFramework, Netcoreapp, Message);
+            this.Setup();
 
             if (IntegrationTestEnvironment.BuildConfiguration.Equals("release", StringComparison.OrdinalIgnoreCase))
             {
